+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This is the log of notable changes to the ustart-cli package.
 
 ### 🎉 New features
 
+- Added 'project name' argument to *init* command
+
 ### 🐛 Bug fixes
 
 ## 2.0